Pass seqs returned from processors through `distinct`.
Without this, the newly added `nondistinct-test` would barf deep
within the internals of skyscraper.traverse, which make the assumption
that the same element won't be conjoined to `:doing` more than
once. If that assumption is violated, `done` hits the unexpected case.

I'm hesitant to fix this at the traverse level because of the
unpredictable performance impact this might incur.