Make core cache backends implement Closeable
Sometimes, the cache backends will need to perform cleanup upon
termination of scraping (e.g., to close a database connection).
To this end, we'll now require backends to implement the
java.io.Closeable interface.

This commits adds implementations of `close` to all the cache
backends in Skyscraper proper. Incidentally, they are all no-ops.