More resilient charset detection in parse-string
Skyscraper was getting confused when the server returned multiple
Content-Type headers (in theory this doesn’t make sense, but it turns
out there are servers in the wild that do that). In this case,
clj-http’s response map will contain a vector under Content-Type.

I really wanted to write a test for this, but it seems it’s not
possible with the current infrastructure due to a limitation of Java
servlets. Ring calls the servlet response’s setContentType() method as
needed, and that method doesn’t do what we want when called twice.