Propagate scraping options through all contexts
This is an internal key in the context, called
:skyscraper.core/options, propagated normally to child contexts but
dissoc'd in the output. This is needed because I'm about to add
an API that will need to know what the options are, given
the context only.

With this, many internal functions that currently take context
and options may be rewritten to accept just context. But that
can wait for later.