Add a helper function to build enhancers
The contract for skyscraper.traverse's :enhancer is that it should
repeatedly read data from one channel and write to another. It almost
should be possible to pass just the context->context enhancing
function, and make `traverse` responsible for setting up the enhancer
loop; however, this isn't quite sufficient for skyscraper.db, since
it needs to wrap the loop in a DB transaction.

So let's not change the contract. Instead, we factor out the
get-enhance-put-loop (GEPL?) into a helper function, and rewrite
skyscraper.db/enhancer with that.